Worked through a couple of more days of the practiceMakesRegex. These were about anchors and groupings.
One of the exercises, it wanted to match la twice in a row and find the
first two letters of that word.  I think it has to be a two step regex.
It'll be interesting to see how the instructor solves it.  I'm thinking in
the case that the word is "lala"  I think there's a way to do it in one
line, but not with the things he's taught us so far.