Just finished the 19th video.
There's one thing that's happening that I'm not sure about.  the
coordinates are showing up as undefined when I look the database through
mongoDB compass.

I also need to look into making an image from the maps info.  Right now
the id I have doesn't allow that.  I get this message when accessing the
url: The Google Maps API server rejected your request. This API project is
not authorized to use this API. Please ensure this API is activated in the
Google Developers Console:
https://console.developers.google.com/apis/api/static_maps_backend?project=_