went through a few more viceos on the python OOP class.
Had an assignment that created a class that overrode the dict class and
stored the values of the dict in a file.  Just had to override the
__init__ function to load the values from the file and store them in the
instance (self) and then update __setitem__ so that it adds the item to
the dict in the instance and then saves the values of the instance to a
file since otherwise you might have multiples of the same key in the saved
file.